linkedinOutreach/requirements.txt
Django==4.0.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2-binary
requests==2.26.0
selenium
gspread
oauth2client
